+++ segmentstore/server/host/src/main/java/io/pravega/segmentstore/server/host/security/TLSConfigChangeHandler.java
@@ -1,4 +1,14 @@
 /**
+ * Copyright (c) 2017 Dell Inc., or its subsidiaries. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+/**
  * Copyright (c) Dell Inc., or its subsidiaries. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -7,7 +17,7 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.pravega.segmentstore.server.security;
+package io.pravega.segmentstore.server.host.security;
 
 import io.netty.handler.ssl.SslContext;
 import lombok.NonNull;
