+++ segmentstore/server/host/src/test/java/io/pravega/segmentstore/server/host/security/TLSConfigChangeEventConsumerTests.java
@@ -1,4 +1,14 @@
 /**
+ * Copyright (c) 2017 Dell Inc., or its subsidiaries. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+/**
  * Copyright (c) Dell Inc., or its subsidiaries. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -7,39 +17,37 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.pravega.segmentstore.server.security;
+package io.pravega.segmentstore.server.host.security;
 
 import io.netty.handler.ssl.SslContext;
 import io.pravega.test.common.SecurityConfigDefaults;
 import org.junit.Test;
-
 import java.io.File;
+import java.nio.file.WatchEvent;
 import java.util.concurrent.atomic.AtomicReference;
 
 import static org.junit.Assert.assertEquals;
+import static org.mockito.Mockito.mock;
 
-public class TLSConfigChangeFileConsumerTests {
+public class TLSConfigChangeEventConsumerTests {
 
-    @Test(expected = NullPointerException.class)
+    @Test (expected = NullPointerException.class)
     public void testNullCtorArgumentsAreRejected() {
-        new TLSConfigChangeFileConsumer(new AtomicReference<>(null), null, null);
+        new TLSConfigChangeEventConsumer(new AtomicReference<>(null), null, null);
     }
 
     @Test (expected = IllegalArgumentException.class)
     public void testEmptyPathToCertificateFileIsRejected() {
-        TLSConfigChangeFileConsumer subjectUnderTest = new TLSConfigChangeFileConsumer(new AtomicReference<>(null),
-                    "", "non-existent");
+        TLSConfigChangeEventConsumer subjectUnderTest = new TLSConfigChangeEventConsumer(new AtomicReference<>(null),
+                "", "non-existent");
         subjectUnderTest.accept(null);
-
-        assertEquals(1, subjectUnderTest.getNumOfConfigChangesSinceStart());
     }
 
     @Test (expected = IllegalArgumentException.class)
     public void testEmptyPathToKeyFileIsRejected() {
-        TLSConfigChangeFileConsumer subjectUnderTest = new TLSConfigChangeFileConsumer(new AtomicReference<>(null),
+        TLSConfigChangeEventConsumer subjectUnderTest = new TLSConfigChangeEventConsumer(new AtomicReference<>(null),
                 "non-existent", "");
         subjectUnderTest.accept(null);
-        assertEquals(1, subjectUnderTest.getNumOfConfigChangesSinceStart());
     }
 
     @Test
@@ -50,13 +58,13 @@ public class TLSConfigChangeFileConsumerTests {
         AtomicReference<SslContext> sslCtx = new AtomicReference<>(TLSHelper.newServerSslContext(
                 new File(pathToCertificateFile), new File(pathToKeyFile)));
 
-        TLSConfigChangeFileConsumer subjectUnderTest = new TLSConfigChangeFileConsumer(sslCtx, pathToCertificateFile,
+        TLSConfigChangeEventConsumer subjectUnderTest = new TLSConfigChangeEventConsumer(sslCtx, pathToCertificateFile,
                 pathToKeyFile);
         subjectUnderTest.accept(null);
 
         assertEquals(1, subjectUnderTest.getNumOfConfigChangesSinceStart());
 
-        subjectUnderTest.accept(null);
+        subjectUnderTest.accept(mock(WatchEvent.class));
         assertEquals(2, subjectUnderTest.getNumOfConfigChangesSinceStart());
     }
 }
