+++ segmentstore/server/host/src/test/java/io/pravega/segmentstore/server/host/security/TLSConfigChangeFileConsumerTests.java
@@ -1,4 +1,14 @@
 /**
+ * Copyright (c) 2017 Dell Inc., or its subsidiaries. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+/**
  * Copyright (c) Dell Inc., or its subsidiaries. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -7,37 +17,39 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.pravega.segmentstore.server.security;
+package io.pravega.segmentstore.server.host.security;
 
 import io.netty.handler.ssl.SslContext;
 import io.pravega.test.common.SecurityConfigDefaults;
 import org.junit.Test;
+
 import java.io.File;
-import java.nio.file.WatchEvent;
 import java.util.concurrent.atomic.AtomicReference;
 
 import static org.junit.Assert.assertEquals;
-import static org.mockito.Mockito.mock;
 
-public class TLSConfigChangeEventConsumerTests {
+public class TLSConfigChangeFileConsumerTests {
 
-    @Test (expected = NullPointerException.class)
+    @Test(expected = NullPointerException.class)
     public void testNullCtorArgumentsAreRejected() {
-        new TLSConfigChangeEventConsumer(new AtomicReference<>(null), null, null);
+        new TLSConfigChangeFileConsumer(new AtomicReference<>(null), null, null);
     }
 
     @Test (expected = IllegalArgumentException.class)
     public void testEmptyPathToCertificateFileIsRejected() {
-        TLSConfigChangeEventConsumer subjectUnderTest = new TLSConfigChangeEventConsumer(new AtomicReference<>(null),
-                "", "non-existent");
+        TLSConfigChangeFileConsumer subjectUnderTest = new TLSConfigChangeFileConsumer(new AtomicReference<>(null),
+                    "", "non-existent");
         subjectUnderTest.accept(null);
+
+        assertEquals(1, subjectUnderTest.getNumOfConfigChangesSinceStart());
     }
 
     @Test (expected = IllegalArgumentException.class)
     public void testEmptyPathToKeyFileIsRejected() {
-        TLSConfigChangeEventConsumer subjectUnderTest = new TLSConfigChangeEventConsumer(new AtomicReference<>(null),
+        TLSConfigChangeFileConsumer subjectUnderTest = new TLSConfigChangeFileConsumer(new AtomicReference<>(null),
                 "non-existent", "");
         subjectUnderTest.accept(null);
+        assertEquals(1, subjectUnderTest.getNumOfConfigChangesSinceStart());
     }
 
     @Test
@@ -48,13 +60,13 @@ public class TLSConfigChangeEventConsumerTests {
         AtomicReference<SslContext> sslCtx = new AtomicReference<>(TLSHelper.newServerSslContext(
                 new File(pathToCertificateFile), new File(pathToKeyFile)));
 
-        TLSConfigChangeEventConsumer subjectUnderTest = new TLSConfigChangeEventConsumer(sslCtx, pathToCertificateFile,
+        TLSConfigChangeFileConsumer subjectUnderTest = new TLSConfigChangeFileConsumer(sslCtx, pathToCertificateFile,
                 pathToKeyFile);
         subjectUnderTest.accept(null);
 
         assertEquals(1, subjectUnderTest.getNumOfConfigChangesSinceStart());
 
-        subjectUnderTest.accept(mock(WatchEvent.class));
+        subjectUnderTest.accept(null);
         assertEquals(2, subjectUnderTest.getNumOfConfigChangesSinceStart());
     }
 }
